regen podcheck after MPE port removed
Commit b5afd3466ff5e5b70ea2921169f138f02727183e removed the MPE/iX
port, which included a file that had known (to podcheck.t) pod errors
podcheck doesn't currently check if a pod has been removed, so its data
base still contained the offending entry, which would be removed the
next time it is regenerated for another unrelated reason.  This might
confuse spelunkers of perl history, so I'm regenerating it separately
now.