avoid overriding prototypes when the call can be fixed
We don't need to call share with an ampersand.  We can just call it in
the way its prototype implies.  While we're at it, we can fix the
prototype that we use when threads::shared can't be loaded.  Also,
there's no need to fake up lock, because it's already a do-nothing weak
keyword without threads.